+++ cypress/integration/workflows_spec.js
@@ -9,17 +9,14 @@ describe('Dashboard Workflows Page', () => {
   });
 
   describe('When logged in', () => {
+    before(() => cy.visit('/'));
     beforeEach(() => {
+      cy.task('resetState');
       cy.login();
-      cy.task('resetState');
-    });
-
-    after(() => {
-      cy.task('resetState');
+      cy.visit('/');
     });
 
     it('displays a link to view workflows', () => {
-      cy.visit('/');
 
       cy.contains('nav li a', 'Workflows').as('workflows');
       cy.get('@workflows').should('have.attr', 'href', '#/workflows');
@@ -28,13 +25,12 @@ describe('Dashboard Workflows Page', () => {
       cy.url().should('include', 'workflows');
       cy.contains('.heading--xlarge', 'Workflows');
 
-      cy.get('table tbody tr').its('length').should('be.eq', 3);
+      cy.get('table tbody tr').its('length').should('be.eq', 2);
       cy.contains('table tbody tr a', 'HelloWorldWorkflow')
         .should('have.attr', 'href', '#/workflows/workflow/HelloWorldWorkflow');
-      cy.contains('table tbody tr a', 'EcsHelloWorldWorkflow')
-        .should('have.attr', 'href', '#/workflows/workflow/EcsHelloWorldWorkflow');
-      cy.contains('table tbody tr a', 'KinesisTriggerTest')
-        .should('have.attr', 'href', '#/workflows/workflow/KinesisTriggerTest');
+      cy.contains('table tbody tr a', 'SecondTestWorkflow')
+        .should('have.attr', 'href', '#/workflows/workflow/SecondTestWorkflow');
+
     });
 
     it('displays a link to individual workflow', () => {
@@ -51,7 +47,8 @@ describe('Dashboard Workflows Page', () => {
       cy.contains('.heading--large', workflowName);
       cy.getJsonTextareaValue().then((workflowJson) => {
         expect(workflowJson.name).to.equal(workflowName);
-        expect(workflowJson.definition.States.HelloWorld).to.exist;
+        expect(workflowJson.definition.States.StartStatus)
+          .to.deep.equal({Type: "Task", Resource: "${SfSnsReportLambdaAliasOutput}", Next: "StopStatus"});
       });
     });
   });