+++ CHANGELOG.md
@@ -9,6 +9,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- **CUMULUS-1102**
+  - Adds ability to run dashboard against Cumulus localAPI.
+    - Adds a number of docker-compose commands to be run via `npm run <command>`
+        - `seed-database` - loads data fixtures into a running stack for testing
+        - `start-localstack` - starts necessary backend for cumulus API. LocalStack + Elasticsearch
+        - `stop-localstack` - stops same
+        - `start-cumulusapi` - starts localstack and cumulus localAPI
+        - `stop-cumulusapi` - stops same
+        - `start-dashboard` - starts localstack, cumulus localAPI and dashboard
+        - `stop-dashboard` - stops same
+        - `e2e-tests` - starts starts localstack, cumulus localAPI, dashboard and cypress end to end tests.
+        - `validation-tests` - starts starts localstack, cumulus localAPI, dashboard and validation tests.
+        - `view-docker-logs`- helper to view logs for the currently running docker-compose stack.
+
+
+
+- **CUMULUS-1463**
+  - Add Datepicker Component
+
 - **CUMULUS-1581**
   - Added support Bulk Granule Operations
 
@@ -17,6 +36,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - **CUMULUS-1463**
   - Add Datepicker Component
+
+### Changed
+
+- ** CUMULUS-1102
+  - Integration (cypress) tests and validation tests run in container against local Cumulus API.
+  - Upgrades to node 10.16.3
 
 ### Fixed
 
@@ -33,6 +58,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **CUMULUS-1679**
   - The UI no longer breaks by producing a blank page when the user types in the
     log search box on the granule details page.
+
+### Deleted
+
+- **CUMULUS-1102**
+  - Removes fake-api.js.  The fake-api is removed in favor of running a Cumulus API locally.
+
 
 ## [v1.6.1] - 2019-11-12